* Fixed warnings in VS2012: C4018 (signed/unsigned mismatch) and C4244 (conversion from 'double' to 'float').

+++ cocos/gui/UIListView.cpp
@@ -301,7 +301,8 @@ void UIListView::removeLastItem()
 
 UIWidget* UIListView::getItem(unsigned int index)
 {
-    if ((int)index < 0 || index >= _items->count())
+    // @todo Signed/unsigned mishmash. Index is must be unsigned int always. Yes?
+    if ( ((int)index < 0) || ((int)index >= _items->count()) )
     {
         return nullptr;
     }
